+++ rbx_types/CHANGELOG.md
@@ -1,6 +1,7 @@
 # rbx_types Changelog
 
 ## Unreleased Changes
+* `Ref` can now represent null explicitly via `Ref::none` and `Ref::is_none`.
 
 ## 0.1.0 (2020-02-05)
 * Initial release of types crate, should have rough feature parity with rbx_dom_weak.